Move ExpandWhens to its own file
